+++ spec/models/upload_management_consultancy_rm6187_spec.rb
@@ -142,14 +142,14 @@ RSpec.describe Upload do
               supplier_framework_lot_services: [],
               supplier_framework_lot_jurisdictions: [
                 {
-                  jurisdiction_id: 'GB'
+                  jurisdiction_id: 'RM6187.GB'
                 }
               ],
               supplier_framework_lot_rates: [
                 {
                   position_id: 'RM6187.1.1',
                   rate: 10000,
-                  jurisdiction_id: 'GB'
+                  jurisdiction_id: 'RM6187.GB'
                 }
               ],
               supplier_framework_lot_branches: []
@@ -160,14 +160,14 @@ RSpec.describe Upload do
               supplier_framework_lot_services: [],
               supplier_framework_lot_jurisdictions: [
                 {
-                  jurisdiction_id: 'GB'
+                  jurisdiction_id: 'RM6187.GB'
                 }
               ],
               supplier_framework_lot_rates: [
                 {
                   position_id: 'RM6187.3.1',
                   rate: 20000,
-                  jurisdiction_id: 'GB'
+                  jurisdiction_id: 'RM6187.GB'
                 }
               ],
               supplier_framework_lot_branches: []
@@ -187,7 +187,7 @@ RSpec.describe Upload do
         it 'assigns attributes to the jurisdictions' do
           result
 
-          expect(supplier_framework.lots.map { |lot| lot.jurisdictions.pluck(:jurisdiction_id) }).to eq([['GB'], ['GB']])
+          expect(supplier_framework.lots.map { |lot| lot.jurisdictions.pluck(:jurisdiction_id) }).to eq([['RM6187.GB'], ['RM6187.GB']])
         end
 
         it 'assigns attributes to the rates' do
@@ -301,14 +301,14 @@ RSpec.describe Upload do
             supplier_framework_lot_services: [],
             supplier_framework_lot_jurisdictions: [
               {
-                jurisdiction_id: 'GB'
+                jurisdiction_id: 'RM6187.GB'
               }
             ],
             supplier_framework_lot_rates: [
               {
                 position_id: 'RM6187.1.88',
                 rate: 10000,
-                jurisdiction_id: 'GB'
+                jurisdiction_id: 'RM6187.GB'
               }
             ],
             supplier_framework_lot_branches: []
@@ -319,14 +319,14 @@ RSpec.describe Upload do
             supplier_framework_lot_services: [],
             supplier_framework_lot_jurisdictions: [
               {
-                jurisdiction_id: 'GB'
+                jurisdiction_id: 'RM6187.GB'
               }
             ],
             supplier_framework_lot_rates: [
               {
                 position_id: 'RM6187.3.1',
                 rate: 20000,
-                jurisdiction_id: 'GB'
+                jurisdiction_id: 'RM6187.GB'
               }
             ],
             supplier_framework_lot_branches: []
@@ -364,14 +364,14 @@ RSpec.describe Upload do
           ],
           supplier_framework_lot_jurisdictions: [
             {
-              jurisdiction_id: 'GB'
+              jurisdiction_id: 'RM6187.GB'
             }
           ],
           supplier_framework_lot_rates: [
             {
               position_id: 'RM6187.1.1',
               rate: 10000,
-              jurisdiction_id: 'GB'
+              jurisdiction_id: 'RM6187.GB'
             }
           ],
           supplier_framework_lot_branches: []
@@ -386,14 +386,14 @@ RSpec.describe Upload do
           ],
           supplier_framework_lot_jurisdictions: [
             {
-              jurisdiction_id: 'GB'
+              jurisdiction_id: 'RM6187.GB'
             }
           ],
           supplier_framework_lot_rates: [
             {
               position_id: 'RM6187.2.1',
               rate: 20000,
-              jurisdiction_id: 'GB'
+              jurisdiction_id: 'RM6187.GB'
             }
           ],
           supplier_framework_lot_branches: []